Create individual files for vite plugins
